requirements.txt
# Базовые зависимости для MyBiz
Django==5.2.9
Pillow==12.0.0
python-decouple==3.8
django-ckeditor==6.7.3

# Для продакшена (раскомментировать при необходимости)
# psycopg2-binary==2.9.9
# whitenoise==6.6.0
# gunicorn==21.2.0
